+++ NPPBiaHoi/ucPhuongTien/frmThemChiPhiPhuongTien.cs
@@ -49,7 +49,7 @@ namespace NPPBiaHoi.ucPhuongTien
                 }
                 ChiPhiPhuongTien aChiPhiPhuongTien = new ChiPhiPhuongTien();
                 aChiPhiPhuongTien.Ten = txtTen.Text;
-                aChiPhiPhuongTien.TienPhi = int.Parse(txtSoTien.Text);
+                aChiPhiPhuongTien.TienPhi = float.Parse(txtSoTien.Text);
                 aChiPhiPhuongTien.MieuTa = mmoMieuTa.Text;
                 aChiPhiPhuongTien.ThoiGian = timeThoiGian.DateTime;
